+++ Controls/listActions.ts
@@ -1,0 +1,17 @@
+/**
+ * Библиотека стандартных действий над записями
+ * @library Controls/listActions
+ * @includes Remove Controls/_listActions/Remove
+ * @includes IAction Control/_listActions/interface/IAction
+ * @includes IActionOptions Controls/_listActions/interface/IActionOptions
+ */
+
+import IAction from './_listActions/interface/IAction';
+import IActionOptions from './_listActions/interface/IActionOptions';
+import Remove from './_listActions/Remove';
+
+export {
+    Remove,
+    IAction,
+    IActionOptions
+}
